requirements.txt
cryptography==39.0.1
python-binance==1.0.17
requests==2.28.2
